count down timer

+++ Assets/TeamAssets/TeamScripts/GameManager.cs
@@ -5,7 +5,7 @@ using UnityEngine.SceneManagement;
 
 public class GameManager : MonoBehaviour {
 
-    public float timeLimit;
+    public int timeLimit;
 
     public GameObject gameMenu;
     public GameObject menuHand;
